refactor(observation): move all element observers to element-observation
